requirements.txt
# access
beautifulsoup4
google-api-python-client
google-auth
google-auth-oauthlib

# app
flask
dotenv
requests
gunicorn

# PODCASTFY

# client
typer
pyyaml

# content_generator
langchain_community
langchain_google_genai
langchain_core
langchain

# text_to_speech
pydub

# PROVIDERS
edge_tts
elevenlabs
openai
google-cloud-texttospeech

# CONTENT_PARSER
google.generativeai
pymupdf
youtube_transcript_api

# Overwrite adhoc
pyOpenSSL